+++ cmd/pricing-cli/main_test.go
@@ -4,13 +4,13 @@ import (
 	"os"
 	"testing"
 
-	chassis "github.com/ai8future/chassis-go/v9"
-	"github.com/ai8future/chassis-go/v9/secval"
-	"github.com/ai8future/chassis-go/v9/testkit"
+	chassis "github.com/ai8future/chassis-go/v10"
+	"github.com/ai8future/chassis-go/v10/secval"
+	"github.com/ai8future/chassis-go/v10/testkit"
 )
 
 func TestMain(m *testing.M) {
-	chassis.RequireMajor(9)
+	chassis.RequireMajor(10)
 	os.Exit(m.Run())
 }